requirements.txt
google-genai
python-dotenv
pymongo
sentence-transformers
